fix(experiments): cast baseline count() to int for Dapper materialization
Postgres count() returns bigint (Int64) but CageBaselineItem/GroupBaselineItem declare
AnimalsWithReading as int (Int32), so Dapper could not match a constructor and threw a 500 on the
baseline by-cage/by-group endpoints. Cast count(...)::int in both queries, matching the pattern the
other session queries already use.

+++ src/Modules/Experiments/SISLAB.Modules.Experiments.Application/Projects/Queries/ListBaselineByGroupQuery.cs
@@ -65,7 +65,7 @@ internal sealed class ListBaselineByGroupQueryHandler
             g.name                     AS groupname,
             g.dose_amount              AS doseamount,
             g.dose_unit                AS doseunit,
-            count(l.animal_id)         AS animalswithreading,
+            count(l.animal_id)::int    AS animalswithreading,
             avg(l.value)               AS meanvalue,
             min(l.value)               AS minvalue,
             max(l.value)               AS maxvalue,
